backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.18
pydantic==2.10.4
celery[redis]==5.4.0
redis==5.2.1
biopython==1.84
numpy==1.26.4
scipy==1.14.1
scikit-learn==1.6.0
sse-starlette==2.1.3
pytest==8.3.4
ruff==0.8.6
